# nodejs-express-openapi/openapi.json
{
   "openapi":"3.0.2",
   "info":{
      "title":"FastAPI",
      "version":"0.1.0"
   },
   "paths":{
      "/shops/{shop}/pets/_search":{
         "post":{
            "summary":"Search Pets",
            "operationId":"searchPets",
            "parameters":[
               {
                  "required":true,
                  "schema":{
                     "title":"Shop",
                     "type":"string"
                  },
                  "name":"shop",
                  "in":"path"
               }
            ],
            "responses":{
               "200":{
                  "description":"Successful Response",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/PetHits"
                        }
                     }
                  }
               },
               "422":{
                  "description":"Validation Error",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         }
      },
      "/shops/{shop}/pets/{pet}":{
         "get":{
            "summary":"Get Pet",
            "operationId":"getPet",
            "parameters":[
               {
                  "required":true,
                  "schema":{
                     "title":"Shop",
                     "type":"string"
                  },
                  "name":"shop",
                  "in":"path"
               },
               {
                  "required":true,
                  "schema":{
                     "title":"Pet",
                     "type":"string"
                  },
                  "name":"pet",
                  "in":"path"
               }
            ],
            "responses":{
               "200":{
                  "description":"Successful Response",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/Pet"
                        }
                     }
                  }
               },
               "422":{
                  "description":"Validation Error",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         }
      }
   },
   "components":{
      "schemas":{
         "HTTPValidationError":{
            "title":"HTTPValidationError",
            "type":"object",
            "properties":{
               "detail":{
                  "title":"Detail",
                  "type":"array",
                  "items":{
                     "$ref":"#/components/schemas/ValidationError"
                  }
               }
            }
         },
         "Pet":{
            "title":"Pet",
            "required":[
               "name",
               "age"
            ],
            "type":"object",
            "properties":{
               "name":{
                  "title":"Name",
                  "type":"string"
               },
               "age":{
                  "title":"Age",
                  "type":"integer"
               }
            }
         },
         "PetHits":{
            "title":"PetHits",
            "required":[
               "hits"
            ],
            "type":"object",
            "properties":{
               "hits":{
                  "title":"Hits",
                  "type":"array",
                  "items":{
                     "$ref":"#/components/schemas/Pet"
                  }
               }
            }
         },
         "ValidationError":{
            "title":"ValidationError",
            "required":[
               "loc",
               "msg",
               "type"
            ],
            "type":"object",
            "properties":{
               "loc":{
                  "title":"Location",
                  "type":"array",
                  "items":{
                     "type":"string"
                  }
               },
               "msg":{
                  "title":"Message",
                  "type":"string"
               },
               "type":{
                  "title":"Error Type",
                  "type":"string"
               }
            }
         }
      }
   }
}